requirements.txt
MarkupSafe==1.1.1
DBUtils==2.0
requests==2.18.1
redis==3.5.3
pyyaml==5.3.1
tornado==6.1
